Reduce wards duplication between transformers
