+++ transformers/storage/cat/repository.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/utilities/wards"
 	"github.com/makerdao/vulcanizedb/libraries/shared/storage"
 	"github.com/makerdao/vulcanizedb/pkg/datastore/postgres"
 )
@@ -13,7 +14,6 @@ const (
 	InsertCatIlkChopQuery = `INSERT INTO maker.cat_ilk_chop (diff_id, header_id, ilk_id, chop) VALUES ($1, $2, $3, $4) ON CONFLICT DO NOTHING`
 	InsertCatIlkFlipQuery = `INSERT INTO maker.cat_ilk_flip (diff_id, header_id, ilk_id, flip) VALUES ($1, $2, $3, $4) ON CONFLICT DO NOTHING`
 	InsertCatIlkLumpQuery = `INSERT INTO maker.cat_ilk_lump (diff_id, header_id, ilk_id, lump) VALUES ($1, $2, $3, $4) ON CONFLICT DO NOTHING`
-	insertWardsQuery      = `INSERT INTO maker.wards (diff_id, header_id, address_id, usr, wards) VALUES ($1, $2, $3, $4, $5) ON CONFLICT DO NOTHING`
 
 	insertCatLiveQuery = `INSERT INTO maker.cat_live (diff_id, header_id, live) VALUES ($1, $2, $3) ON CONFLICT DO NOTHING`
 	insertCatVatQuery  = `INSERT INTO maker.cat_vat (diff_id, header_id, vat) VALUES ($1, $2, $3) ON CONFLICT DO NOTHING`
@@ -33,8 +33,8 @@ func (repository *CatStorageRepository) Create(diffID, headerID int64, metadata 
 		return repository.insertVat(diffID, headerID, value.(string))
 	case Vow:
 		return repository.insertVow(diffID, headerID, value.(string))
-	case Wards:
-		return repository.insertWards(diffID, headerID, metadata, value.(string))
+	case wards.Wards:
+		return wards.InsertWards(diffID, headerID, metadata, repository.ContractAddress, value.(string), repository.db)
 	case IlkChop:
 		return repository.insertIlkChop(diffID, headerID, metadata, value.(string))
 	case IlkFlip:
@@ -63,46 +63,6 @@ func (repository *CatStorageRepository) insertVat(diffID, headerID int64, vat st
 func (repository *CatStorageRepository) insertVow(diffID, headerID int64, vow string) error {
 	_, writeErr := repository.db.Exec(insertCatVowQuery, diffID, headerID, vow)
 	return writeErr
-}
-
-func (repository *CatStorageRepository) insertWards(diffID, headerID int64, metadata storage.ValueMetadata, wards string) error {
-	user, userErr := getUser(metadata.Keys)
-	if userErr != nil {
-		return userErr
-	}
-
-	tx, txErr := repository.db.Beginx()
-	if txErr != nil {
-		return txErr
-	}
-
-	addressID, addressErr := shared.GetOrCreateAddress(repository.ContractAddress, repository.db)
-	if addressErr != nil {
-		rollbackErr := tx.Rollback()
-		if rollbackErr != nil {
-			return shared.FormatRollbackError("cat address", addressErr.Error())
-		}
-		return addressErr
-	}
-
-	userAddressID, userAddressErr := shared.GetOrCreateAddress(user, repository.db)
-	if userAddressErr != nil {
-		rollbackErr := tx.Rollback()
-		if rollbackErr != nil {
-			return shared.FormatRollbackError("cat user address", userAddressErr.Error())
-		}
-		return addressErr
-	}
-
-	_, insertErr := tx.Exec(insertWardsQuery, diffID, headerID, addressID, userAddressID, wards)
-	if insertErr != nil {
-		rollbackErr := tx.Rollback()
-		if rollbackErr != nil {
-			return shared.FormatRollbackError("cat wards with address", insertErr.Error())
-		}
-		return insertErr
-	}
-	return tx.Commit()
 }
 
 // Ilks mapping: bytes32 => flip address; chop (ray), lump (wad) uint256
@@ -160,12 +120,4 @@ func getIlk(keys map[storage.Key]string) (string, error) {
 		return "", storage.ErrMetadataMalformed{MissingData: constants.Ilk}
 	}
 	return ilk, nil
-}
-
-func getUser(keys map[storage.Key]string) (string, error) {
-	user, ok := keys[constants.User]
-	if !ok {
-		return "", storage.ErrMetadataMalformed{MissingData: constants.User}
-	}
-	return user, nil
 }
