+++ transformers/storage/vow/keys_loader_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/test_helpers"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/utilities/wards"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/vow"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/test_data"
 	"github.com/makerdao/vulcanizedb/libraries/shared/factories/storage"
@@ -90,9 +91,9 @@ var _ = Describe("Vow storage keys loader", func() {
 			wardsUser := fakes.FakeAddress.Hex()
 			storageRepository.WardsKeys = []string{wardsUser}
 			paddedWardsUser := "0x000000000000000000000000" + wardsUser[2:]
-			wardsKey := common.BytesToHash(crypto.Keccak256(common.FromHex(paddedWardsUser + vow.WardsMappingIndex)))
+			wardsKey := common.BytesToHash(crypto.Keccak256(common.FromHex(paddedWardsUser + wards.WardsMappingIndex)))
 			expectedMetadata := vdbStorage.ValueMetadata{
-				Name: vow.Wards,
+				Name: wards.Wards,
 				Keys: map[vdbStorage.Key]string{constants.User: fakes.FakeAddress.Hex()},
 				Type: vdbStorage.Uint256,
 			}
